Walls Collision
Reworked the movement of the player.

Added collision on the tilemap and the player.
